refactor: move ContainerOptions alongside Container
