+++ src/index.ts
@@ -1,6 +1,5 @@
-export type { Container } from "./container";
+export type { Container, ContainerOptions } from "./container";
 export { createContainer } from "./container";
-export type { ContainerOptions } from "./containerOptions";
 export { AutoRegister, Inject, Injectable, InjectAll, Scoped } from "./decorators";
 export type { ClassDecorator, ClassFieldDecorator } from "./decorators/decorators";
 export { inject, injectBy } from "./inject";
